Change int32 to int64 input tensors in the test cases.

+++ tests/test_afi.py
@@ -5,7 +5,7 @@ from fm_zoo.afi import AutomaticFeatureInteraction
 
 
 def test_afi():
-    x = tf.convert_to_tensor([[1, 2, 2, 1, 5], [0, 3, 2, 2, 4], [2, 2, 2, 1, 1]])
+    x = tf.convert_to_tensor([[1, 2, 2, 1, 5], [0, 3, 2, 2, 4], [2, 2, 2, 1, 1]], dtype="int64")
     m = AutomaticFeatureInteraction([3, 4, 5, 4, 6], 4, 2, 2, [5, 3, 1], .1)
     o = m(x)
